just custom json encoding

+++ schema/collections.capnp
@@ -3,6 +3,6 @@
 
 struct UiViewMetadata {
   title @0 :Text;
-  dateSaved @1 :Float64; # milliseconds since unix epoch
+  dateSaved @1 :UInt64; # milliseconds since unix epoch
   addedBy @2 :Text; # Identity ID, encoded in hexadecimal format.
 }
